backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
pymupdf
python-multipart
python-dotenv
